Move Dp, Density, and Margin into better package
