+++ redwood-layout-composeui/src/androidUnitTest/kotlin/app/cash/redwood/layout/composeui/ComposeUiFlexContainerTest.kt
@@ -31,7 +31,7 @@ import app.cash.redwood.flexbox.JustifyContent
 import app.cash.redwood.layout.AbstractFlexContainerTest
 import app.cash.redwood.layout.TestFlexContainer
 import app.cash.redwood.layout.api.Constraint
-import app.cash.redwood.layout.api.Margin
+import app.cash.redwood.ui.Margin
 import app.cash.redwood.widget.Widget
 import org.junit.Rule
 import org.junit.runner.RunWith
