+++ redwood-layout-composeui/src/androidUnitTest/kotlin/app/cash/redwood/layout/composeui/ComposeUiSpacerTest.kt
@@ -23,8 +23,8 @@ import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
 import app.cash.redwood.LayoutModifier
 import app.cash.redwood.layout.AbstractSpacerTest
-import app.cash.redwood.layout.api.dp
 import app.cash.redwood.layout.widget.Spacer
+import app.cash.redwood.ui.dp
 import app.cash.redwood.widget.Widget
 import com.android.ide.common.rendering.api.SessionParams
 import org.junit.Rule
