+++ redwood-layout-dom/src/main/kotlin/app/cash/redwood/layout/dom/utils.kt
@@ -17,10 +17,10 @@ package app.cash.redwood.layout.dom
 
 import app.cash.redwood.layout.api.Constraint
 import app.cash.redwood.layout.api.CrossAxisAlignment
-import app.cash.redwood.layout.api.Density
-import app.cash.redwood.layout.api.Dp
 import app.cash.redwood.layout.api.MainAxisAlignment
 import app.cash.redwood.layout.api.Overflow
+import app.cash.redwood.ui.Density
+import app.cash.redwood.ui.Dp
 import kotlin.math.roundToInt
 
 internal fun Dp.toPxString(): String = with(Density(1.0)) {
