+++ redwood-layout-view/src/test/kotlin/app/cash/redwood/layout/view/ViewSpacerTest.kt
@@ -24,8 +24,8 @@ import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
 import app.cash.redwood.LayoutModifier
 import app.cash.redwood.layout.AbstractSpacerTest
-import app.cash.redwood.layout.api.dp
 import app.cash.redwood.layout.widget.Spacer
+import app.cash.redwood.ui.dp
 import app.cash.redwood.widget.Widget
 import com.android.ide.common.rendering.api.SessionParams
 import org.junit.Rule
