+++ redwood-runtime/src/jvmMain/kotlin/app/cash/redwood/ui/density.kt
@@ -13,18 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package app.cash.redwood.layout.api
+package app.cash.redwood.ui
 
-import android.content.res.Resources
-
-// Android uses 2.75 as a density scale for most recent Pixel devices and iOS
-// uses 3. This aligns the two so the generic values used by Redwood layout are
-// visually similar on both platforms.
-internal actual const val DensityMultiplier = 1.1
-
-/**
- * Create a [Density] from the device's [Resources].
- */
-public fun Density(resources: Resources): Density {
-  return Density(resources.displayMetrics.density.toDouble())
-}
+internal actual const val DensityMultiplier = 1.0
